core/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Question, Answer

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username')

class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = '__all__'

class AnswerSerializer(serializers.ModelSerializer):
    user = UserSerializer(read_only=True)
    total_likes = serializers.ReadOnlyField()

    class Meta:
        model = Answer
        fields = '__all__'